// src/app/pages/model/discharge.ts
export class Discharge {
    baby_id = "";
    ventilation = false;
    non_invasive = false;
    surfactant = false;
    nasal_prong = false;
    inotropes = false;
    jaundice = false;
    phototherapy = false;
    day_phototherapy: number = null;
    max_bilirubin: number = null;
    exchange_transfusion = false;
    is_blood_culture = false;
    blood_culture_result: Array<any> = [];
    blood_culture_other_g_plus = "";
    blood_culture_other_g_minus = "";
    blood_culture_other_fungal = "";
    blood_culture_other = "";
    is_other_blood_culture = false;
    other_blood_culture_type: number = null;
    other_blood_culture_type_other = "";
    other_blood_culture_result: Array<any> = [];
    other_blood_culture_g_plus = "";
    other_blood_culture_g_minus = "";
    other_blood_culture_fungal = "";
    other_blood_culture_other = "";
    is_antibiotics = false;
    antibiotics_type: Array<any> = [];
    antibiotics_type_other = "";
    cooling = false;
    cooling_start_before = false;
    cooling_temperature: number = null;
    cooling_start: any = new Date();
    rewarming_start: any = new Date();
    cooling_stop: any = new Date();
    cerebral_ultrasound = false;
    ivh_grade: number = null;
    is_pvl = false;
    is_hydrocephalus = false;
    cerebral_other = "";
    cardiac_echo = false;
    cardiac_echo_result = "";
    blood_products = false;
    parenteral_nutrition = false;
    full_milk_feeding: any = new Date();
    central_line = false;
    discharge: any = new Date();
    principal_diagnosis: any = null;
    other_diagnosis: Array<any> = [];
    outcome: number = 2;
    is_intensive = false;
    dead_cause: number = null;
    feeding_at_discharge = false;
    transfer_to: number = null;
    transfer_summary = "";
    transfer_fio2: number = null;
    transfer_respiration_support: number = null;
    transfer_cpap_pressure: number = null;
    transfer_peak_pressure: number = null;
    transfer_peep: number = null;
    transfer_ventilator_rate: number = null;
    transfer_tube_size: number = null;
    transfer_last_blood_gas: any = new Date();
    transfer_last_full_blood: any = new Date();
    transfer_last_full_blood_result = "";
    transfer_last_electrolytes: any = new Date();
    transfer_other_result = "";
    transfer_current_antibiotics = "";
    transfer_current_drugs = "";
    transfer_positive_culture = "";
    transfer_fluid_volume: number = null;
    transfer_intravenous_fluid = false;
    transfer_fluid_type = "";
    transfer_intravenous_fluid_volume: number = null;
    transfer_milk = false;
    transfer_milk_types: Array<number> = [];
    transfer_milk_other = "";
    transfer_milk_volume: number = null;
    transfer_blood_gas: any;
    transfer_blood_gas_date:any;
    transfer_blood_gas_ph: any;
    transfer_blood_gas_pco2: any;
    transfer_blood_gas_co2: any;
    transfer_blood_gas_base
    transfer_full_blood: any;
    transfer_full_blood_hb: any;
    transfer_full_blood_wcc: any;
    transfer_full_blood_newtro: any;
    transfer_full_blood_plate: any;
    transfer_electrolytes: any;
    transfer_electrolytes_sodium: any;
    transfer_electrolytes_potassium: any;
    transfer_electrolytes_urea: any;
    transfer_electrolytes_calcium: any;
    transfer_electrolytes_creatinine
}